fix: replace deprecated datetime.utcnow() with datetime.now(timezone.utc)
- Updated game_lifecycle_service.py to use timezone-aware datetime
- Updated notification_service.py to use timezone-aware datetime
- Fixed test mocks in test_game_lifecycle_service.py
- Fixed datetime usage in test_notification_service.py
- Added return type annotations to websocket_manager.py methods
- Fixed apply_handicap_strokes return type in rule_manager.py

This resolves deprecation warnings for Python 3.12+.

+++ backend/app/managers/rule_manager.py
@@ -889,7 +889,7 @@ class RuleManager:
         hole_number: int,
         player_handicaps: Dict[str, float],
         hole_stroke_index: int
-    ) -> Dict[str, int]:
+    ) -> Dict[str, float]:
         """
         Calculate handicap strokes for each player on a specific hole.
